Refactor: Move _first_call(), _second_call(), and _LlmService__third_call() from llm_service.py to llm_executor.py
